Update 16/09 18:29

+++ Lab01/bai3.py
@@ -1,4 +1,3 @@
 def triple(num):
-    return (num * 3)
-
+    return int(num * 3)
 print(triple(3))
